requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
abseil-cpp=20230802.0=h6a678d5_2
absl-py=2.0.0=pyhd8ed1ab_0
aiohttp=3.9.1=py311h459d7ec_0
aiosignal=1.3.1=pyhd8ed1ab_0
annotated-types=0.7.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
blas=1.0=mkl
blinker=1.7.0=pyhd8ed1ab_0
bottleneck=1.3.7=py311hf4808d0_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotli-python=1.0.9=py311h6a678d5_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.24.0=hd590300_0
ca-certificates=2024.7.4=hbcca054_0
cachetools=5.3.2=pyhd8ed1ab_0
certifi=2024.7.4=pyhd8ed1ab_0
cffi=1.16.0=py311h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=unix_pyh707e725_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=py311hdb19cb5_0
cryptography=41.0.7=py311hdda0065_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.99=0
cuda-runtime=12.1.0=0
cudatoolkit=11.8.0=h6a678d5_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py311h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py311h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=py311h459d7ec_0
fsspec=2023.10.0=py311h06a4308_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.13=h85f3911_1
google-auth=2.25.2=pyhca7485f_0
google-auth-oauthlib=1.2.0=pyhd8ed1ab_0
grpc-cpp=1.48.2=he1ff14a_4
grpcio=1.48.2=py311he1ff14a_4
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
gtest=1.14.0=hdb19cb5_0
huggingface-hub=0.27.0=pypi_0
icu=73.1=h6a678d5_0
idna=3.4=py311h06a4308_0
importlib-metadata=7.0.0=pyha770c72_0
importlib_metadata=7.0.0=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=8.26.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=py311h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py311h38be061_0
kiwisolver=1.4.4=py311h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h166bdaf_1003
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcblas=3.9.0=5_h92ddd45_netlib
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.0.20=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.5.119=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libexpat=2.5.0=hcb278e6_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_3
libgfortran-ng=13.2.0=h69a702a_5
libgfortran5=13.2.0=ha4646dd_5
libgomp=13.2.0=h807b86a_3
libhwloc=2.7.1=h32351e8_0
libiconv=1.17=hd590300_2
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=5_h92ddd45_netlib
libllvm14=14.0.6=hdb19cb5_3
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libopenblas=0.3.26=pthreads_h413a1c8_0
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libuv=1.48.0=hd590300_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
llvm-openmp=12.0.1=h4bd325d_1
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.1=pyhd8ed1ab_0
markupsafe=2.1.1=py311h5eee18b_0
matplotlib=3.3.4=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
matplotlib2tikz=0.7.6=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
multidict=6.0.4=py311h459d7ec_1
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
networkx=3.1=py311h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numexpr=2.8.7=py311h65dcdc2_0
numpy=1.26.2=py311h08b1b3b_0
numpy-base=1.26.2=py311hf175353_0
oauthlib=3.2.2=pyhd8ed1ab_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.3.1=h4bc722e_2
packaging=23.1=py311h06a4308_0
pandas=2.2.1=py311ha02d727_0
parso=0.8.4=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.0.1=py311ha6cbd5a_0
pip=23.3.1=py311h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
ply=3.11=py311h06a4308_0
prompt-toolkit=3.0.47=pyha770c72_0
protobuf=3.20.3=py311h6a678d5_0
psutil=5.9.8=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyasn1=0.5.1=pyhd8ed1ab_0
pyasn1-modules=0.3.0=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.10.3=pypi_0
pydantic-core=2.27.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_0
pyopenssl=23.2.0=py311h06a4308_0
pyparsing=3.0.9=py311h06a4308_0
pyqt=5.15.10=py311h6a678d5_0
pyqt5-sip=12.13.0=py311h5eee18b_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.5=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.11=2_cp311
pytorch=2.2.1=py3.11_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py311h06a4308_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0.1=py311h459d7ec_1
pyzmq=25.1.2=py311h6a678d5_0
qt-main=5.15.2=h53bd1ea_10
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.31.0=py311h06a4308_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
safetensors=0.4.5=pypi_0
seaborn=0.12.2=py311h06a4308_0
sentry-sdk=2.19.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py311h06a4308_0
sip=6.7.12=py311h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h9b69904_2
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.15.1=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py311h63ff55d_1
tensorboardx=2.2=pyhd3eb1b0_0
tikzplotlib=0.10.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.21.0=pypi_0
torchtriton=2.2.0=py311
torchvision=0.17.1=py311_cu121
tornado=6.3.3=py311h5eee18b_0
tqdm=4.65.0=py311h92b7b1e_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.47.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.18=py311h06a4308_0
wandb=0.19.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=24.11.1=pypi_0
werkzeug=3.0.1=pyhd8ed1ab_0
wheel=0.41.2=py311h06a4308_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.4.5=h5eee18b_0
yaml=0.2.5=h7f98852_2
yarl=1.9.3=py311h459d7ec_0
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
